+++ src/components/Compact/style/index.js
@@ -10,6 +10,8 @@ export const controllerPrefix = prefixCls + '-controller';
 export const CompactWrap = styled.div.attrs({
     className: prefixCls
 })`
+    position: relative;
+    z-index: 1;
     margin-left: 1px;
 
     .${controllerPrefix} {