+++ src/components/Tabs/style/index.js
@@ -12,7 +12,7 @@ export const prefixCls = _prefixCls + '-tabs';
 export const TabsWrap = styled(RcTabs).attrs({
     className: ({ styleType }) => `${prefixCls}-styletype-${styleType}`
 })(
-    ({ theme: { colorMap, colorList, Tabs: tabsTheme }, styleType, tabBarPosition }) => css`
+    ({ theme: { colorMap, colorList, Tabs: tabsTheme }, tabBarPosition }) => css`
         overflow: hidden;
         ${clearFixMixin};
 
@@ -81,8 +81,8 @@ export const TabsWrap = styled(RcTabs).attrs({
                 color: ${colorMap.disabled.text};
             }
 
-            ${styleType === 'default' &&
-                css`
+            &-styleType-default-bar {
+                .${prefixCls} {
                     &-ink-bar {
                         display: none !important;
                     }
@@ -94,10 +94,11 @@ export const TabsWrap = styled(RcTabs).attrs({
                     &-tab-active:hover {
                         ${tabsTheme['default:active']};
                     }
-                `};
+                }
+            }
 
-            ${styleType === 'ink' &&
-                css`
+            &-styleType-ink-bar {
+                .${prefixCls} {
                     &-ink-bar {
                         width: 2px;
                         height: 2px;
@@ -111,7 +112,8 @@ export const TabsWrap = styled(RcTabs).attrs({
                     &-tab-active {
                         ${tabsTheme['ink:active']};
                     }
-                `};
+                }
+            }
         }
 
         &.${prefixCls} {
@@ -121,10 +123,9 @@ export const TabsWrap = styled(RcTabs).attrs({
                         border-bottom: 1px solid ${colorList.secondary4};
                         .${prefixCls} {
                             &-nav-container {
-                                margin-bottom: -1px;
+                                bottom: -1px;
                             }
                             &-tab {
-                                margin-bottom: -1px;
                                 display: inline-block;
 
                                 &:hover {
@@ -151,10 +152,9 @@ export const TabsWrap = styled(RcTabs).attrs({
                         border-top: 1px solid ${colorList.secondary4};
                         .${prefixCls} {
                             &-nav-container {
-                                margin-top: -1px;
+                                top: -1px;
                             }
                             &-tab {
-                                margin-top: -1px;
                                 display: inline-block;
 
                                 &:hover {
@@ -183,11 +183,9 @@ export const TabsWrap = styled(RcTabs).attrs({
                         height: 100%;
                         .${prefixCls} {
                             &-nav-container {
-                                margin-right: -1px;
+                                right: -1px;
                             }
                             &-tab {
-                                margin-right: -1px;
-
                                 &:hover {
                                     border-right-color: ${colorList.secondary4};
                                 }
@@ -215,11 +213,9 @@ export const TabsWrap = styled(RcTabs).attrs({
                         height: 100%;
                         .${prefixCls} {
                             &-nav-container {
-                                margin-left: -1px;
+                                left: -1px;
                             }
                             &-tab {
-                                margin-left: -1px;
-
                                 &:hover {
                                     border-left-color: ${colorList.secondary4};
                                 }